make getUsersInterests function
